+++ src/engineio/transports/polling.rs
@@ -1,4 +1,4 @@
-use crate::engineio::transport_new::Transport;
+use crate::engineio::transport::Transport;
 use crate::error::{Error, Result};
 use bytes::{BufMut, Bytes, BytesMut};
 use native_tls::TlsConnector;
